refactor: name the lifted modules the way the repo names modules
The convention is `<feature>.<type>.ts`, where the dotted segment names a KIND of
module and the subject uses hyphens — `window-emptiness.util.ts`,
`owner-routed-command.util.ts`, `target-shard-resolver.util.ts`. Three of the four
new files put a subject where the type belongs, and one of them spelled the same
feature two ways in two directories, since `web-view-move.model.ts` already
existed. `shard-index` is the exception: it already reads as a kind, and its
sibling and dependency `service-shard-index.ts` is bare and hyphenated, so it
matches that.

The standard that sanctions this lift names three exemplars to copy, and all three
are hyphenated-subject `.util.ts`. Renamed with `git mv`, so history follows.

Two corrections in the same files, from the review of the disposition change.

The comment above that throw, and the message it throws, both still described the
old `not-reopened` meaning. The message had become self-contradictory — it said
the window is holding the content and that the view "was not reopened anywhere" in
one sentence — and that is the text that lands in the log, for the reader who most
needs an accurate account. Changing the token and leaving its prose behind is the
same failure the change was fixing, one file over.

A comment in the fold-in also wrapped mid-phrase, orphaning a word across two
lines.
